GroupManageController RESTful API 적용
1. View담당 GroupManageViewController 생성
2. 데이터 담당 GroupManageController클래스 RestController로 수정후
URL수정
3. 그에 따른 JS파일 수정
